+++ integrations/OpenGL/FragmentShaderOpenGL.cpp
@@ -53,11 +53,11 @@ bool FragmentShader::initialize()
 		// then functions for each expression in the shader; we can iterate the sorted
 		// expressions and put out definitions for each unique expression ClassDef
 		std::set<const Reflection::ClassDef*> uniqueExprs;
-		int exprCount = mMaterial->expressionCount();
+        size_t exprCount = mMaterial->expressionCount();
 		MaterialExpression** expressions = mMaterial->sortedExpressions();
 
 		std::stringstream defs;
-		for (int i = 0; i<exprCount; ++i) {
+        for (size_t i = 0; i<exprCount; ++i) {
 			MaterialExpression* expr = expressions[i];
 			const Reflection::ClassDef* classDef = expr->getDerivedClassDef();
 			if (uniqueExprs.find(classDef) == uniqueExprs.end()) {
@@ -66,8 +66,8 @@ bool FragmentShader::initialize()
 				// and then generate the definition for this expression
 				if (expr->getDerivedClassDef() == Sampler2DExpression::getClassDef()) {
 					Sampler2DExpression* sampExp = static_cast<Sampler2DExpression*>(expr);
-					int texIndex = 0;
-					int sampIndex = 0;
+                    size_t texIndex = 0;
+                    size_t sampIndex = 0;
 					sampExp->appendDefinition(MaterialExpression::SHADER_GLSL4, texIndex, sampIndex, defs);
 				} else {
 					expr->appendDefinition(MaterialExpression::SHADER_GLSL4, defs);
@@ -98,7 +98,7 @@ bool FragmentShader::initialize()
 		AttrToVarName names;
 
 		// go through all connections and form this map
-		int nConnections = mMaterial->connections(0, 0);
+        size_t nConnections = mMaterial->connections(0, 0);
 
 		//  there may not be connections yet, so do the following only if connections exist
 		if (nConnections) {
@@ -108,8 +108,8 @@ bool FragmentShader::initialize()
 			// variable names should be based on the name (type) of the expression
 			// that owns the "from" (output) attr; for uniqueness, we'll just assign
 			// monotonically-increasing ordinals to each connection
-			int ord = 0;
-			for (int i = 0; i < nConnections; ++i) {
+            size_t ord = 0;
+            for (size_t i = 0; i < nConnections; ++i) {
 				std::stringstream ss;
 				Gfx::Attribute* out = connections[i]->output();
 				MaterialExpression* me = out->mParent;
@@ -123,7 +123,7 @@ bool FragmentShader::initialize()
 
 		// then generate the function calls...
 		std::stringstream calls;
-		for (int i = 0; i<exprCount; ++i) {
+        for (size_t i = 0; i<exprCount; ++i) {
 			MaterialExpression* expr = expressions[i];
 
 			// our input attributes are someone else's output attributes, so collect their names
@@ -143,7 +143,7 @@ bool FragmentShader::initialize()
 				}
 			}
 
-			int nameIdx = 0;
+            size_t nameIdx = 0;
 			for (size_t j = 0; j<inputAttrs.size(); ++j) {
 				// find this attribute's name in the connection map
 				const Attribute* attr = &inputAttrs[j];
