+++ tech/Gfx/VertexElement.h
@@ -54,19 +54,19 @@ namespace Teardrop
             VertexElement();
 
 			// vertex element size in bytes
-			int size();
+            size_t size();
 
 			// element data math type
 			VertexElementType mType;
 
 			// number of elements (usually [1..4])
-			int mCount;
+            size_t mCount;
 
 			// what this element is meant to be used for 
 			VertexElementUsage mUsage;
 
 			// the zero-based Nth instance of this element in the vertex (i.e. 2 for TEXCOORD2)
-			int mIndex;
+            size_t mIndex;
 
 			TD_DECLARE_ALLOCATOR();
 		};
